Clean up typing modules
