+++ frontend/src/types/error.ts
@@ -1,0 +1,3 @@
+import { RequestError } from '@octokit/types';
+
+export type PluginRepoFetchError = Pick<RequestError, 'name' | 'status'>;
